+++ src/clip/AudioClipState.ts
@@ -1,0 +1,10 @@
+'use strict'
+
+/**
+ * Describes the different states available for an audio clip.
+ */
+export enum AudioClipState {
+  Stopped = "STOPPED",
+  Playing = "PLAYING",
+  Paused = "PAUSED"
+}