+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+0.0.2 / 2020-01-29
+==================
+* [FEATURE] Added the following methods to the base API: `getAudio`, `removeAudio`, `removeAllAudio`.]
+* [FEATURE] Added AudioClipState enum to prepare for adding pause/resume functionality next update.
+* [DOCS] Added new methods to README.
+
 0.0.1 / 2020-01-28
 ==================
 * Initial commit